requirement.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
langchain-text-splitters
langchain-openai
beautifulsoup4
langchain-ollama
faiss-cpu
langchain-groq
fastapi
uvicorn
langserve
sse_starlette
langchain-chroma
bs4
streamlit
wikipedia
arxiv
google
ddgs